// src/models/product_example_content_type.ts
import { IContentItem, Elements } from '@kentico/kontent-delivery';

/**
 * Generated by '@kentico/kontent-model-generator@4.1.0'
 */
export type ProductExampleContentType = IContentItem<{
  url: Elements.UrlSlugElement;
  description: Elements.RichTextElement;
  name: Elements.TextElement;
  image: Elements.AssetsElement;
  taxonomyCategorizingTargetAudience: Elements.TaxonomyElement;
}>;